Remove MD5, significantly clean up code, update documentation
The way fdupes works means that a small number of hash collisions
are acceptable and are handled properly. Jody Bruchon's hashing
function is significantly faster and computationally very simple,
but MD5 is overkill in this application. A lot of code was left in
place to continue supporting MD5, but in all tests the new way is
faster with no errors. MD5 is therefore being thrown out.

Lots of "#if 0" and commented-out code blocks were removed. Some
inconsistent coding styles regarding if/else braces were fixed up.

Documentation was updated to reflect new changes, contributors,
and to remove a few TODO items that are already fulfilled.