More intelligent numeric sort, small file speedup
The numeric_sort() function would sort numbers correctly, but the
comparison would fall back on naive byte value comparisons when a
digit sequence was not directly being examined. This resulted in
files with Windows-style "copy but keep both files" names being in
the opposite order from what would normally be desired. A deletion
of such duplicates would look like this:

[+] file001 (1).jpg
[-] file001.jpg
[-] file001-1.jpg

When it obviously should be deleting the copy version instead:

[+] file001.jpg
[-] file001-1.jpg
[-] file001 (1).jpg

numeric_sort() was modified to force some symbols and spaces to
always sort after alphanumeric characters, solving this problem
and bringing the auto-deletion order in line with what a normal
user would typically desire and expect.

fdupes performs partial hashing of files to quickly eliminate
obvious non-matches without reading the entire file. Unfortunately
it was performing the hash calculation twice for files that were
at or below the size of PARTIAL_HASH_SIZE. In such cases, the
hash of the "partial" file is always the hash of the entire file
as well. Before running two hash calculations plus an extra hash
comparison, the file sizes are checked and the redundant work is
avoided. This can result in a significant speedup when processing
huge amounts of small files (1 to 4096 bytes long).