Optimize hard link comparison when -H option is passed
The flag F_CONSIDERHARDLINKS (toggled ON with the -H option) was
simply bypassing the hard link check which skips hard linked pairs
of files. This meant that hard linked files would be compared as if
they were only possible duplicates. Hard linked files are always
duplicates which means there is no need to compare the files. This
change skips all comparisons on hard linked files and forces them
to be matched without touching the file contents.