Update license information, add license to 'fdupes -v'
The license requires that its text be included in all substantial
portions of the software. Added the license text to the version
information output. Scripts that scrape 'fdupes -v' for version
checking should pipe through 'head -n 1' to ignore the license
text output.

Also updated copyright range to extend to 2015.