Change getopt behavior, escapefilename() tweak, NO_UID_GID changes
The getopt string was being conditionally generated. This could
get very complex, plus it would be better to ignore parameters
which would be valid if missing capabilities were present so as to
avoid breaking scripts that run fdupes. Thus, getopt was restored
to using one single option string. Options that are not valid in a
specific compilation of fdupes will simply be ignored.

escapefilename() was using malloc() for its temporary variable and
not free()ing it or reusing it later. Switched escapefilename() to
a static char array and added bounds checking to accommodate it.

NO_UID_GID was added because UNIX permissions are not properly
supported under MinGW (and Windows in general). The file mode was
not being considered part of this because it did not stop the
compilation of the program, but file mode changes seem to not have
any effect on Windows and permission bits are mostly meaningless.
Changed to NO_PERMS and conditionally removed all permisssion code
when compiling for Windows.