Add dupe pair count to progress, change some ints to longs
It is nice to know how effective the duplicate scan is while the scan is
in progress. This change shows the number of matched file pairs beside
the file counter in the progress indicator.

Some of the counters used in main() for tracking file and progress counts
were changed to 'long' types. These ideally should be moved to uint64_t
explicitly in the near future. While it is unlikely that anyone will be
using fdupes with more than 2097151 files in the set, it is possible and
the program needs to take this into consideration to avoid odd behavior
on integer overflows.