Add the dangerous, data-destroying -Q/--quick option
For large data sets where a small amount of data loss is
acceptable to the user, it may be better to speed up the duplicate
scan rather than performing byte-by-byte file comparisons for all
match pairs. This option relies solely on hashes of files and will
effectively avoid reading any complete file more than once.

The tradeoff is that hash collisions make this dangerous for the
purpose of not losing non-duplicate files. Because of the danger
involved, the option is only documented in the manual page and no
mention of it should ever appear in the text for the -h/--help
options.