Add -O option to sort result sets by command line order
fdupes has always suffered from barely-controllable behavior in
produced result sets. The only significant control so far has been
using the -o option to change the pair sort algorithm. As part of
my effort to have more control over the result set output, this
commit adds a -O/--paramorder option to force fdupes to sort all
results in the order their parent directories were specified on
the command line and fall back to normal sorting only for pairs
that are in the same specified parent directory.

This option allows the necessary control over the combined -d and
-N options which automatically delete all but the first result in
a duplicate set without prompting. Without this option, fdupes
can order result sets in ways that are undesirable for the user
and leaves the user with no way to delete duplicates in the order
they actually prefer short of manual deletion or writing a shell
script wrapper around fdupes.

To illustrate the effect of this option, this is an (abbreviated)
output in a linux-4.1 source directory which has been compiled and
has lots of redundant built-in.o files to play with.

$ ./fdupes -n auxdisplay/ net/team/ firewire/
auxdisplay/built-in.o
firewire/built-in.o
net/team/built-in.o

$ ./fdupes -nO auxdisplay/ net/team/ firewire/
auxdisplay/built-in.o
net/team/built-in.o
firewire/built-in.o

Without the new option, fdupes retains its original behavior.