Load progress improvement, use larger ints for future-proofing
File list loading was just a spinning pipe which gives the user
no real information about the actual loading progress. Switched
to showing the number of file entries loaded to provide somewhat
more meaningful feedback to the user. This should be modified to
have the number of command-line directory specifications processed
to be even more informative.

There was some inconsistency in the sizing of integers used for
counting files and duplicates. The smallest type used was a signed
int which is limited to 2,147,483,647 on most modern systems.
While this is certainly a large number, there are already systems
out there which have more files than this. To future-proof fdupes
against being run on such huge file sets, the int and long types
for file counting were switched to uintmax_t.

This commit also removes two unused variables from the hard link
code: 'group' and 'curgroup'. They were incremented in loops but
were never actually used for anything.