Further improve loading progress indicator, fix some printf()s
When fdupes loads the list of files to be checked, progress now
shows the number of files and directories scanned. Additionally,
the number of specified directories from the command line that
have been processed appears after these counts. This provides far
more meaningful loading progress than a rotating pipe or simple
file count.

Some printf() statements weren't properly converted for the use of
uintmax_t and needed to be cleaned up. A cast which was no longer
necessary was removed.

The width of the progress clearing statement was expanded due to
the extra information now shown in duplicate scanning sometimes
being longer than 40 characters.