Make hard linking (-L option) much safer, update CHANGES
Hard linking used to be performed by unlink() and then link() on a
file. This is dangerous because the destruction of a file occurs
whether or not the hard link succeeds. On filesystems that do not
support hard linking such as VFAT this makes the -L option result
in loss of all but the first file in a duplicate set. Now the code
renames the file with a temporary extension, then links, then when
all of those operations succeed it deletes the renamed original.
If an error is encountered with hard linking then the file is
renamed back to its original name.