Move hash initialization so it stops clobbering the hash value
In the getcrcsignatureuntil() file read loop, the hash was being
reset to zero on every chunk read, effectively making the hash
computed and returned only run over the last chunk of data read
from the file. This probably caused a lot of unnecessary file
comparisons when the start and end chunks of two files were the
same but the contents in between were not.

Testing against a compiled Linux 4.1 kernel source tree shows an
improvement on that particular data set of about 3.3%, however
the real-world change could be faster, particularly in situations
where the system doesn't have as much memory as my development
machines do (typically 8GB) and the data set is large enough to
blow away parts of itself from the system's disk caches.