Numerous hard link improvements
Hard linking did not make sure that files resided on the same
device before attempting a hard link, resulting in annoying error
messages. Revised code to check for different devices and simply
warn the user that hard linking is not possible across devices.
Note that this will prevent any files not on the same device as
the source files to not be linked amongst themselves, but that was
already the behavior anyway.

Previously, the -L and -H options were marked as not compatible
and using both would result in an error message and program exit.
This behavior has been changed; now the hard link code checks if
a pair is already hard linked and skips over such pairs, taking no
action. The practical use of this is to retrieve a list of files
that are hard linked already. Hard linked pairs show a different
prefix marker that indicates they are equal: -==->

The fdupes visual cues for file deletion are offset from the start
of the line and sort of ugly. While I would like to clean them up,
the need arose for a better output form for hard linking due to
the many possible "no-link" conditions that can be encountered.
When hard linking with the -L option, output now indicates the
file that is being used as the hard link source with [SRC] and
various arrow indicators are prefixed to the remaining files in
the set to indicate the action taken, in the form -xx-> where 'xx'
is the action taken: '--' is "hard linked," '==' is "already hard
linked (no action)," and '//' is "not linking." All -//-> arrows
are printed to stderr and are accompanied by an explanation why
the warning was issued.

The [f]printf() usage in the hard link section could use some
rethinking and cleanup.