Remove redundant getfilestats() call and unneeded goto statement
When getfilestats() was added, the initial implementation of the
function was put into the other functions by search-and-replace
because there were so many different functions that ultimately
led to nothing more than returning one element from a stat() call.

getfilestats() works by performing only one stat() for every file
being considered and simply returning without taking action if it
is called a second time for the same file. This effectively made
the second call near the end of checkmatch() a no-op that would
do nothing more than waste a little bit of time calling and then
immediately returning. This is probably an insignificant boost to
overall performance, but every little bit helps.

While looking at the previous redundant call, it also became clear
that the goto for hard link matches could be replaced with the
same return that it jumps to. If debugging stats are enabled, this
particular goto was also placed in such a way that it would throw
partial_to_full off (noticing two hard links and auto-matching
wouldn't qualify for the stat since it has nothing to do with the
hashes in the first place.)