Refactor checkmatch() hard link checking code
When the code was changed to automatically consider hard links as
matched pairs if the -H flag is passed, the inode and device check
was duplicated in the code. Since these checks can't be skipped in
either case (-H or not -H) there is no reason to check the flag
and then have two separate copies of the inode/device checks also.

Instead, we now check for hard link status in all cases and the
-H flag determines whether we return NULL or a pair match.