Don't compare hard links at all when using -H
checkmatch() was skipping hard link comparisons as it should be.
Unfortunately, the code was calling confirmmatch() for hard links
which meant that hard linked files were byte-for-byte compared
despite being skipped over for partial and full hashing. This
commit corrects this behavior by checking for hard links before
confirming a match the long way. The -H flag is checked before
the hard link status is checked to avoid minor cache thrashing
effects that could be encountered unnecessarily when -H is off.