Eliminate a strdup()/free() by reusing a static temporary variable
If -A is used, strdup() invokes malloc() and free() for every file
encountered. This reduces performance and is also not necessary
since the static temporary variable 'tempname' is free to re-use
at this point in the code. This commit changes the code to take
advantage of tempname and eliminate malloc() calls.