Update copyright years, add feature flags, fix documentation
Copyright years were extended to 2016.

Documentation was updated for the fdupes -> fdupes-jody change.

fdupes-jody now outputs feature flags in the version text. This
makes it easier to know what compile-time changes exist in the
version you're running, and also enables shell scripts to detect
if a particular feature is installed. The recently added btrfs
'-B/--dedupe' switches don't work without btrfs support explicitly
compiled in, so if a shell script wants to use it, it can detect
the feature independently from an actual runtime error.