Add TOCTTOU protection to hard linking code
Hard linking now performs "time of check to time of use" checks
on file pairs before attempting to hard link them. This prevents a
file in a pair that was modified after the pair was matched from
being processed as a duplicate. If the hard link source file is
the file that got modified, the current file becomes the new hard
link source file and the linking loop starts over with the file
after that one in the duplicate chain. This behavior is much safer
than assuming that all files have remained untouched.