Workaround for >1023 hard links on Windows platforms
This change provides a workaround for the 1023-link limit in
Windows. When hard linking files, all files that currently have
a 1024-link count (1 original + 1023 hard links) are removed from
match consideration completely. Each file in the candidate pair is
checked prior to hard linking to ensure its link count has not hit
the link limit. If the destination file has hit the limit, it is
skipped. If the source file used for linking has hit the limit, it
is dropped as the source file and the current candidate becomes
the new source file.

This creates multiple groups of hard links but it is the best we
can possibly do on Windows. The biggest advantage is that the
source file hitting the link limit will not prevent further
linking from happening and the cryptic "error: no error" message
goes away.