Handle hard link problems when permissions might be in the way
On Windows in particular, if the ACL for a file doesn't allow the
user to delete the file but does allow renaming, we tend to find
that the file is moved to a temporary file for hard linking and
the hard linking succeeds, but then the temporary file cannot be
removed due to the ACL. If this happens, we shouldn't leave things
in a messy state with a hard linked version and the temporary name
version side by side. This commit adds cleanup code that will try
to reverse the hard linking process if such a failure occurs.