Improve Windows hard link support, eliminate getino.c/.h entirely
Windows provides a 64-bit file ID that we can use to detect hard
links. Unfortunately, the getino.c routine pulled from GnuWin32
decimates this 64-bit ID into a 16-bit hash which could cause some
files to be detected as hard links even if they're not due to hash
collisions. There is no reason we can't use the 64-bit file ID
as the "inode number." A caveat is that MinGW's sys/types.h says
an ino_t is an "unsigned short" so we have to manually use a
uint64_t to take advantage of the wider type. We now typedef a new
jdupes_ino_t depending on whether or not we're using MinGW so that
Windows uses uint64_t while non-Windows uses ino_t.

This code discards all of the getino.c/.h code in favor of merging
some variables with the get_nlink() code and directly calling
GetFileInformationByHandle() to retrieve the necessary values.

getino() was renamed get_file_id() and the parameters for calling
get_file_id() and get_nlink() have been changed to only require
a pointer to the filename rather than an entire 'struct file_t'.

A nice side effect of all of this is that the new get_file_id() is
greatly simplified when compared to getino() and should be much
faster.