Place Windows stat()-like functions out into their own .c/.h files
This merges get_file_id and get_nlink into an external module that
provides a more generic function called win_stat() which performs
the GetFileInformationByHandle() call and returns information that
is similar to stat() but is Windows-specific. It also includes
macros similar to those used with stat() modes to read various
file attribute flags more easily.

This commit also fixes a bug introduced in v1.1 where grokdir()
would always recurse as if -r was set, even if it wasn't set.