Only speed up progress if we did "long work" like full file reads
The progress indicatior is delayed primarily to avoid slowdown
related to waiting on printf() statements to finish thousands of
times for a file set. This delay was too long for sets of many
large files, so a speedup based on the size of the file being
processed was implemented. This did not take into account whether
the entire file was actually read at any point.

This commit adds a flag variable to tell the progress indicator
that the file being processed did "long work" which means that at
least one file had to be read in its entirety. This causes updates
to progress to only be sped up if the program spent a lot of time
doing I/O, but relieves the pressure in the case where the files
are failing the quick exclusion criteria. The reduced updates
significantly improve performance with lots of large files that
fail to match early in the matching process.