Add CTRL+C handler and -Z/--hardabort option
If you abort jdupes with CTRL+C it will now process the duplicates
it found before the scan was canceled. If the signal to abort is
received during directory enumeration or after duplicate scanning
is complete, the program will terminate immediately. This is very
helpful when aborting a lengthy duplicate scan because the work
done is not completely lost.