Tweak chunk size, add static/restrict/etc. to some elements
Cachegrind says that if the chunk size is too large it'll cause
huge cache misses. This change makes chunk size externally tunable
and reduces the default to 64K which drops jody_hash cache misses
from ~9M to ~355K data read misses on a real-world data set and
a machine with a Phenom II X6 1035T processor.