v1.5: Invert meaning of -Z switch, --hardabort -> --softabort
The CTRL+C handler behavior could cause unsuspecting users that
move up to versions with the new "soft abort" feature to try to
abort program execution to prevent actions from being taken and
discover that the actions get taken anyway. The feature mode was
inverted and the name changed to reflect the new usage. This will
require users to explicitly ask for a soft abort on CTRL+C to use
the feature and reduce the risk of nasty surprises.

Normally this would be moved to a new switch, but the -Z feature
was only available for about six weeks. It's unlikely that it will
cause more pain to switch the meaning than the feature's original
behavior could cause.