Fix BTRFS dedupe code, make ENABLE_BTRFS the preferred option
BTRFS dedupe was broken because the files to be deduplicated must
be opened read-write, not read-only. This change fixes the issue
and adds some loud debug points and code comments to that code.

The original way to add BTRFS support was to use the define
-DHAVE_BTRFS_IOCTL_H which was clumsy and turned out to not be the
correct header name anyway. The preferred way to enable BTRFS code
in jdupes is now to pass ENABLE_BTRFS=1 to 'make' or to add the
define -DENABLE_BTRFS to the CFLAGS. The old way will remain
supported just in case it's needed; it may be removed later but
for now it simply acts as a synonym for ENABLE_BTRFS.