Implement read-only open() fallback for BTRFS dedupe
This currently doesn't seem to work. It's probably in the kernel
and even the duperemove tool by the BTRFS people doesn't seem to
be able to do dedupe of a read-only filesystem anyway. It's there
now, so if the kernel allows read-only opens to be deduped in the
future we'll be ahead of the curve.