string_malloc: support allocations larger than SMA page size
The string_malloc allocator works by allocating a fixed page size
and then handing pieces of those pages out as requested. If a
memory allocation request is larger than the page size, the malloc
would simply fail. This adds a pass-through call to malloc() if
a request for a large object is made. This shouldn't be happening
because the purpose of string_malloc is to make many small allocs
a lot faster and large objects shouldn't be allocated with it, but
it's better to be more robust when possible.