Add LOW_MEMORY compilation option for reduced memory usage
This doesn't reduce memory usage by a whole lot, but it might be
helpful to someone on a platform with low memory availability. On
a test file set with over 200,000 files, this option reduced the
peak memory allocation from 71 MB to about 65 MB. It may also be
faster when on 32-bit platforms or using CPUs with very small
caches since the hash size is reduced to 32 bits.