Change per-file status flags to bit flags, update documentation
Five integers were used as status flags for each file to indicate
if a hash was set, if a file was a symlink, and other purposes. By
collapsing them into a single 'flags' variable, the program runs
faster and also uses less memory. A test run showed a 5% reduction
in memory usage and a very small but consistent speed increase.
On systems with small CPU caches the speed improvement will be
even better.