Prevent attempts to dedupe more than 65535 files
The kernel's dedupe count is an unsigned 16-bit integer which
limits the number of dedupes passed to 65535. If someone really
needs to be able to dedupe a larger file set than this, they'll
see the new error message and should ask for the feature to be
coded into jdupes. If no one needs it, I don't see the point in
increasing the complexity of the dedupe code to support it, but
an unsigned integer overflow here would be bad so the condition
needs to be checked.