Track directories to avoid duplicate traversals
This change adds a tree of device:inode pairs for each directory
traversed to avoid traversing the same directory twice. This is
often seen when someone references the same path twice on the
command line. Before this change, specifying a path twice would
cause all files to match themselves, resulting in potentially very
destructive behavior.

This incurs a small performance penalty but the improvement to
overall safety is worth the cost.