Add -0/--zeromatch, deprecate -n/--noempty, update manual page
The default behavior of considering zero-length files to be
duplicates can be dangerous, especially with the hard and soft
link actions jdupes has introduced. Novices who didn't know about
the zero-length = duplicate default behavior and to use -n to
override it stood a greater chance of accidental data loss.

This commit discards the -n option and adds the -0 option instead.
Now the default behavior is as if -n is always specified and the
-0 option overrides this.

The manual page was updated to reflect current options correctly.