Remove MIT+GPL dual licensing in some code; update copyright info
The MIT License is GPL-compatible. There is no need to offer the
GPL as a licensing option if MIT is offered, so the GPL option has
been removed to simplify everything. All code in jdupes is now
covered under The MIT License exclusively.

Copyright info in some files (specifically headers) was removed to
make any changes simpler in the future.

Copyright years in copyright notices were extended to cover 2017.

Some TODO items were removed; they were already fixed but the TODO
document was not updated. It may be removed soon. Bugs should be
documented in README instead.