Change progress to use gettimeofday() instead of awkward counters
The progress delay has been difficult to keep user-friendly from
the start. Using gettimeofday() gives the user an update every
one second during most operations. The progress will still stick
when scanning large files, but that can be improved in the future.
For now, this is WAY better, accomplishing the goals that the
original delay mechanisms were made for while not compromising the
user experience. The progress delay code is now simpler as well.