Fix a major bug in string_malloc()'s free list functionality
A counter was being used in a place where it shouldn't have been
which resulted in items not being removed from the free list when
reused and some memory getting wrecked. There were also some
confusing uses of code to calculate in the extra space needed for
the size prefix; some of these were probably buggy as well. All
of these problems seem to be fixed and the crashes reported in
Github issues #33 and #34 should now be fixed as a result.