Fix stderr redirect on Windows, use C99 printf specifiers
On Windows, stderr redirection was causing UTF-16 to be mixed with
ANSI/UTF-8 output. It turns out that MSVCRT doesn't output any
text at all if stderr is set to _O_U8TEXT mode. This fix does two
things to correct behavior of redirected stderr on Windows: it
changes the mode of stderr if it's not a TTY (previously only
stdout was being changed) and the non-TTY mode of choice has been
changed to _O_BINARY instead of _O_U8TEXT so that the internal
UTF-8 strings that jdupes works with anyway can just be dumped
straight to stderr without needing any fancy handling.

This commit also changes the use of the intmax_t %jd/%ju printf
specifiers to use C99 specifiers PRIdMAX/PRIuMAX instead. The use
of the -pedantic compiler warning flag was also removed from the
default flags in the Makefile.

This closes one of the problems reports in Github issue #37.