Update documentation and help text
The new -X/--exclude option was not properly documented and the
-x/--xsize option (which is now deprecated) has been rewritten as
a wrapper that uses the new -X code. The available size suffixes
have been significantly enhanced to include base 10 multipliers
and are valid for -x and -X alike. The -X code removes a TODO from
the list. The -@, -D, and -Q options now have help text available.