Make -H option work for individual files too
Accepting individual files instead of only directories did not
respect the -H option because the same double-traversal code for
directories was being used on files. This change does not add any
regular files to the double-traversal avoidance tree which would
have the side effect of forcing -H off for all file arguments.