Do not output progress indicators if stderr is not a TTY
POSIX.1-2008 says that stderr should only be used if the program
exit status will not be "success," so jdupes should generally not
be sending anything to stderr unless it's an actual error.
However, stderr is the only logical place to output progress and
other informational data for user feedback without polluting
stdout; there is no good alternative. As a compromise, this commit
makes jdupes detect if stderr is not a TTY and enables the -q
option to disable progress indicator output when that is the case.

This should fix bug #87 reported by eribertomota where the Debian
Continuous Integration framework fails jdupes because of stderr
output being non-empty when redirected to a file.