Add -M/--printwithsummary option (-m/--summarize + print matches)
It has been requested in passing several times that the ability to
print matches AND print a summary of the matches be made available
in jdupes. At the request of Saso <x64000@gmail.com> this commit
adds the option to combine these two features. Since printing the
match sets has no explicit option and is simply the default action
if no other actions are specified, a new option has been added for
this combination (-M or --printwithsummary).