Fix bug that makes deprecated -x/--xsize option not always work
Accidentally used strcat() instead of strcpy() which resulted in
random failures if the memory allocated for the temporary string
had something in it that also constituted a string. In the best
case (null byte as first byte) this was not a problem, however if
any non-zero byte(s) were present this made the string passed to
add_exclude() invalid, and in the absolute worst case it could
trigger a segfault.