Update and clarify help text and documentation
jdupes can take inidividual files as arguments now, and the
default behavior was not clearly documented in the help text or
documentation. This commit updates the documentation to fix these
problems.

Thanks to /u/ObnoxiousOldBastard on Reddit for reporting the lack
of documentation for default behavior and inspiring this patch!