Support PROGRAM_NAME in 'make install', add 'make stripped'
The person who added PROGRAM_NAME support to the Makefile missed
the 'install' section. This fixes that problem.

A new target 'stripped' was added which will strip the executable
after building. The Windows executable in particular tends to be
massively bloated with debugging stuff after building and when it
gets stripped it's 1/10 the size! Adding a stripping target will
make it just a bit faster to put together binary packages.