Add build date to version info screen, make -V a synonym for -v
When doing test builds, the version number does not change. This
adds the exact build date to the version info screen so that any
different builds can be differentiated by their build dates.

This also makes -V do the same thing as -v. I may remove this in
the future but it could be helpful, so I see no harm in doing it.
-V is unused otherwise so who cares?