Add -T/--partial-only option; fix partial match print behavior
The -T/--partial-only option matches based on the early exclusion
and partial hashing phases ONLY, skipping full hashing and full
file byte-for-byte comparison. Yes, the implication is correct:
this new option ignores all file data beyond the first block which
is usually 4096 bytes of data, making it an extremely dangerous
option when paired with file actions such as deletion or linking.
It is provided for comparison in special situations such as two
downloads that may be damaged or incomplete, or providing a list
of possibly but not definitely identical files to the user very
quickly as a heuristic, allowing the user to take further action
manually.

Because of the extreme danger of data loss that the new option can
cause, it must be specified at least twice to take effect and will
trigger an immediate error if only specified once. Use at your own
risk. You're welcome.

Also, a fix to the '-P partial' option is applied. When files were
passed through the partial-to-full hash optimization code due to
being under the partial hash block size they'd also bypass the
reporting that partial hashes matched. This moves that code to a
place where it will be printed before that optimization decision
is made and should now reliably print small files that match when
the '-P partial' option is used.