Add -t/--no-tocttou option to work around filesystem issues
User u/go-ive on Reddit brought up the following issue:

"jdupes refused to actually remove any duplicates after hours of
scanning. Multiple times. It insisted that the files have been
changed during the scan, which they surely have not."

Network and exotic filesystems sometimes fail to report consistent
information for files when polled multiple times, but the safety
check in jdupes that aborts any actions on changed files will see
such differences as changes that make operations on the changed
file unsafe. This feature is designed to stop so-called TOCTTOU
(time of check to time of use) errors which can pose a security
risk in some cases, but it relies on information about a file not
changing across calls to stat(). Network filesystems in particular
are prone to inconsistencies; CIFS without UNIX extensions and the
'serverino' mount option will change inode numbers, for example:

https://unix.stackexchange.com/q/421680/44505

Because a user report of this issue is now "in the wild" and there
are known peculiarities with certain filesystems that can trigger
an inappropriate TOCTTOU safety warning, this commit adds the -t
option which causes file_has_changed() to always return 0 rather
than performing any file change checks, effectively disabling the
check. This reduces safety and very slightly increases performance
due to a drastic reduction in stat() calls. For data sets where a
user knows no changes should happen, the performance boost may be
preferred over the safety check anyway.