Make exclude stack handle strings safely; add LOUD debug messages
The exclusion stack string_malloc() size was not accounting for
the trailing null byte. There was also an unnecessary strcpy()
call when the source string is an empty string. None of the -X
parameters that are currently implemented are "no-value" types,
but I've been toying with adding some and found the code was not
fully up to handling an exclusion parameter that takes no value.

Also added more LOUD debug messages because there are never ever
enough of those bad boys.