Change --no-tocttou to the more intuitive --nochangecheck
Users are having trouble finding the option to use when they are
dealing with an underlying filesystem that reports inconsistent
file modify time and inode metadata. A long option that actually
self-documents is far preferable, thus "no change check" is the
new long option and the old one is no longer supported. Using -t
is not affected.