Fix a null pointer check in travdone_free()
I copied and pasted the line and didn't rewrite both instances of
"left" to "right." Yes, I felt like an idiot as soon as I noticed.

Fortunately, the only side effect this actually has is that the
free operation will traverse the leftmost side of the tree to its
maximum depth, then free every node along the way back to the top,
which means it won't ever cause a crash, it just doesn't actually
free the majority of the tree.