Fix crash on Windows where long option requires an argument
Apparently getopt_long() on current MSYS2 MinGW compilers is not
working properly; rather than a long option requiring a parameter
returning opt='?'/optarg=NULL, it returns the correct value for
when the option's required argument is provided along with optarg
set to the pointer address 0x20 (on both i386 and x86_64). This
causes improper usage to trigger a segfault instead of a graceful
error message and exit. This commit checks for the bogus optarg
pointer and handles it properly.