Fix dedupe compilation, including automatic header fallback
This includes code in act_dedupefiles.c that will check for the
new Linux 4.5+ FIDEDUPERANGE support and automatically includes
the static support header (with a warning) on any system that is
using the outdated kernel headers. Note that the change from the
BTRFS-specific names to the generalized FIDEDUPERANGE names is
only cosmetic in the kernel, so it really doesn't matter.

To silence the warning, pass STATIC_DEDUPE_H=1 to make.