Implement strtoepoch() to avoid unavailable strptime()
strptime() is not available on Windows. It also requires some
feature test macros that it'd be nice to just not bother with.
This is probably a lot faster than strptime() anyway, though this
is not even remotely a critical path in the code.

This should fix -X newer/older for Windows and work everywhere.