Add -U/--notravcheck for recursion in Google Drive File Stream
GDFS returns the same dev:ino pair for all directories, tripping
the double-traversal protection and eliminating all recursion.
This makes it possible to dupe-scan GDFS drives that do this.

In addition, this adds several LOUD debug messages to options and
moves the getopt string to a more conveniently editable spot.