Make -X size consistent and update documentation
The behavior of the size extfilter and the documentation of its
behavior were not consistent. This makes both the behavior and the
documentation consistent. This changes -X size to be an inclusion
filter rather than an exclusion filter, which could be a problem
for existing shell scripts, but is more consistent with the whole
"filter by this criterion" concept of extfilter. Annoying warnings
have been sprinkled in the code and README to warn users of this
behavioral change, and these warnings will be removed after the
v1.19.0 release.

This also re-folds the README to 80 columns to make things a bit
more visually pleasant.

Fixes #139 and superseded #140