Make mtime removable; add (but disable) atime stuff
This makes all features that rely on storing the file modification
time in memory optional and adds the fundamentals for using file
access times. In the near future, jdupes will be able to touch
files without the file access time changing. For now, since this
feature is not implemented yet, the Makefile disables file access
time storage.

Other small changes include INSTALL corrections and additions,
removal of getfilestats() as an extern function (it's not used in
any other code modules) and un-inlining it since the compiler can
see that it's only used in one place and take correct action.